requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
ultralytics>=8.0.0
paddleocr>=2.7.0
paddlepaddle>=2.5.0
numpy>=1.24.0
pyyaml>=6.0
tqdm>=4.65.0
scikit-image>=0.21.0
